refractor for better readability + pass lint checks
