+++ apps/attendance/app/_components/status-indicator.tsx
