+++ apps/attendance/app/_components/time-display.tsx
