+++ apps/attendance/app/admin/dashboard/_components/user-detail/loading-state.tsx
