+++ apps/attendance/lib/format.ts
