+++ apps/attendance/package.json
@@ -2,11 +2,12 @@
   "name": "spike-attendance",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev --turbopack --port 3005",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint --max-warnings 0"
   },
   "dependencies": {
     "@spike/api": "workspace:*",
@@ -36,6 +37,8 @@
     "zod": "^4.0.15"
   },
   "devDependencies": {
+    "@spike/eslint-config": "workspace:*",
+    "@spike/typescript-config": "workspace:*",
     "@tailwindcss/postcss": "^4",
     "@types/node": "^20",
     "@types/pg": "^8.15.5",
@@ -47,6 +50,5 @@
     "tw-animate-css": "^1.3.6",
     "typescript": "5.9.2",
     "wrangler": "^4.28.1"
-  },
-  "packageManager": "pnpm@10.11.1+sha512.e519b9f7639869dc8d5c3c5dfef73b3f091094b0a006d7317353c72b124e80e1afd429732e28705ad6bfa1ee879c1fce46c128ccebd3192101f43dd67c667912"
+  }
 }
